refactor: Implement unit test (Car.test.js) and car names and number test (carValidation.test.js)
